# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IMPACCT Brooklyn Content Submission Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentBody"
      responses:
        "201":
          description: Department created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /departments/{id}:
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDepartmentBody"
      responses:
        "200":
          description: Department updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /resources:
    get:
      operationId: listResources
      tags: [resources]
      summary: List all resources
      responses:
        "200":
          description: List of resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resource"
    post:
      operationId: createResource
      tags: [resources]
      summary: Add a resource (any authenticated user)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResourceBody"
      responses:
        "201":
          description: Resource created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /resources/{id}:
    patch:
      operationId: updateResource
      tags: [resources]
      summary: Edit a resource (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResourceBody"
      responses:
        "200":
          description: Resource updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteResource
      tags: [resources]
      summary: Archive/delete a resource (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Resource deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /resource-categories/suggest:
    post:
      operationId: suggestResourceCategory
      tags: [resourceCategories]
      summary: Suggest a new resource category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestCategoryBody"
      responses:
        "201":
          description: Category suggestion submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceCategory"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /resource-categories/{id}/approve:
    patch:
      operationId: approveResourceCategory
      tags: [resourceCategories]
      summary: Approve a pending resource category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceCategory"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /resource-categories/{id}:
    delete:
      operationId: deleteResourceCategory
      tags: [resourceCategories]
      summary: Delete a resource category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Category deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Department:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contact_email:
          type: ["string", "null"]
        active:
          type: ["boolean", "null"]
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - created_at
    CreateDepartmentBody:
      type: object
      properties:
        name:
          type: string
        contact_email:
          type: ["string", "null"]
        active:
          type: ["boolean", "null"]
      required:
        - name
    UpdateDepartmentBody:
      type: object
      properties:
        name:
          type: string
        contact_email:
          type: ["string", "null"]
        active:
          type: ["boolean", "null"]
...
    Resource:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category_id:
          type: integer
        description:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        login_info_location:
          type: ["string", "null"]
        access_level:
          type: string
          enum: [everyone, department, leadership]
        cost:
          type: ["string", "null"]
        renewal_date:
          type: ["string", "null"]
          format: date
        added_by:
          type: integer
        notes:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
        category_name:
          type: ["string", "null"]
        added_by_name:
          type: ["string", "null"]
      required:
        - id
        - name
        - category_id
        - access_level
        - added_by
        - created_at
    CreateResourceBody:
      type: object
      properties:
        name:
          type: string
        category_id:
          type: integer
        description:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        login_info_location:
          type: ["string", "null"]
        access_level:
          type: string
          enum: [everyone, department, leadership]
        cost:
          type: ["string", "null"]
        renewal_date:
          type: ["string", "null"]
          format: date
        notes:
          type: ["string", "null"]
      required:
        - name
        - category_id
    UpdateResourceBody:
      type: object
      properties:
        name:
          type: string
        category_id:
          type: integer
        description:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        login_info_location:
          type: ["string", "null"]
        access_level:
          type: string
          enum: [everyone, department, leadership]
        cost:
          type: ["string", "null"]
        renewal_date:
          type: ["string", "null"]
          format: date
        notes:
          type: ["string", "null"]
    ResourceCategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [active, pending]
        suggested_by:
          type: ["integer", "null"]
        approved_by:
          type: ["integer", "null"]
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - created_at
    SuggestCategoryBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name