# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IMPACCT Brooklyn Content Submission Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceAccessLevel } from "./resourceAccessLevel";

export interface Resource {
  id: number;
  name: string;
  category_id: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  login_info_location?: string | null;
  access_level: ResourceAccessLevel;
  /** @nullable */
  cost?: string | null;
  /** @nullable */
  renewal_date?: Date | null;
  added_by: number;
  /** @nullable */
  notes?: string | null;
  created_at: Date;
  /** @nullable */
  category_name?: string | null;
  /** @nullable */
  added_by_name?: string | null;
}