# lib/api-zod/src/generated/types/resourceCategory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IMPACCT Brooklyn Content Submission Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceCategoryStatus } from "./resourceCategoryStatus";

export interface ResourceCategory {
  id: number;
  name: string;
  status: ResourceCategoryStatus;
  /** @nullable */
  suggested_by?: number | null;
  /** @nullable */
  approved_by?: number | null;
  created_at: Date;
}