# lib/api-zod/src/generated/types/updateResourceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IMPACCT Brooklyn Content Submission Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateResourceBodyAccessLevel } from "./updateResourceBodyAccessLevel";

export interface UpdateResourceBody {
  name?: string;
  category_id?: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  login_info_location?: string | null;
  access_level?: UpdateResourceBodyAccessLevel;
  /** @nullable */
  cost?: string | null;
  /** @nullable */
  renewal_date?: Date | null;
  /** @nullable */
  notes?: string | null;
}